// src/app/core/services/model/eTestCaseStatus.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Status of the TwmsBackend.Services.DteFunctionTestService.DteTestCases
 */
export type ETestCaseStatus = 'Ready' | 'Processing' | 'WaitingForInput' | 'Success' | 'Failed' | 'Canceled' | 'Aborted';

export const ETestCaseStatus = {
    Ready: 'Ready' as ETestCaseStatus,
    Processing: 'Processing' as ETestCaseStatus,
    WaitingForInput: 'WaitingForInput' as ETestCaseStatus,
    Success: 'Success' as ETestCaseStatus,
    Failed: 'Failed' as ETestCaseStatus,
    Canceled: 'Canceled' as ETestCaseStatus,
    Aborted: 'Aborted' as ETestCaseStatus
};