// src/app/core/services/model/testSeriesResult.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TestCaseResult } from './testCaseResult';
import { User } from './user';

/**
 * Result of a TwmsBackend.Services.DteFunctionTestService.TestSeries.
 */
export interface TestSeriesResult { 
    testUser?: User;
    startTime?: Date;
    endTime?: Date;
    testId?: string;
    deviceUuid?: string;
    testCaseResults?: Array<TestCaseResult>;
}