// src/app/core/services/model/testStartInput.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Input for starting the test.
 */
export interface TestStartInput { 
    deviceUuid?: string;
    name?: string;
}